refactor: remove v2 separation for simplicity
