+++ inlang/source-code/versioned-interfaces/project-settings/src/index.ts
@@ -1,1 +1,1 @@
-export { ProjectSettingsV2 as ProjectSettings } from "./interface.v2.js"
+export { ProjectSettings } from "./interface.js"
